Move the bordered selection into its own file.
