+++ macos/Bookmarks/Views/BorderedSelection.swift
@@ -1,0 +1,47 @@
+// Copyright (c) 2020-2021 InSeven Limited
+ //
+ // Permission is hereby granted, free of charge, to any person obtaining a copy
+ // of this software and associated documentation files (the "Software"), to deal
+ // in the Software without restriction, including without limitation the rights
+ // to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ // copies of the Software, and to permit persons to whom the Software is
+ // furnished to do so, subject to the following conditions:
+ //
+ // The above copyright notice and this permission notice shall be included in all
+ // copies or substantial portions of the Software.
+ //
+ // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ // FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ // AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ // SOFTWARE.
+
+import SwiftUI
+
+struct BorderedSelection: ViewModifier {
+
+    @Environment(\.applicationHasFocus) var applicationHasFocus
+
+    var selected: Bool
+    var firstResponder: Bool
+
+    var color: Color {
+        applicationHasFocus && firstResponder ? Color.accentColor : Color.unemphasizedSelectedContentBackgroundColor
+    }
+
+    func body(content: Content) -> some View {
+        if selected {
+            content
+                .padding(4)
+                .overlay(
+                    RoundedRectangle(cornerRadius: 13)
+                        .stroke(color, lineWidth: 3))
+        } else {
+            content
+                .padding(4)
+        }
+    }
+
+}
